Moved dataset/wordvectors files to util/wordvectors.
